v1exbackend/pyproject.toml
[tool.poetry]
name = "ai-goals-tracker-backend"
version = "2.0.0"
description = "Backend server for AI Goals Tracker with LangGraph agents and real-time WebSocket communication"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Database
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pgvector = "^0.2.4"

# Redis
redis = {extras = ["hiredis"], version = "^5.0.1"}
aioredis = "^2.0.1"

# RabbitMQ
aio-pika = "^9.3.1"

# LangChain & LangGraph
langchain = "^0.1.0"
langchain-openai = "^0.0.2"
langgraph = "^0.0.20"
langsmith = "^0.0.77"

# OpenAI
openai = "^1.10.0"

# MinIO (S3)
minio = "^7.2.3"
boto3 = "^1.34.29"

# Data formats
pyarrow = "^15.0.0"
pandas = "^2.2.0"

# WebSocket
websockets = "^12.0"
python-socketio = "^5.11.0"

# Utils
python-dotenv = "^1.0.0"
httpx = "^0.26.0"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.0"
ruff = "^0.1.14"
mypy = "^1.8.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
